feat: Implement ability to observe properties on any class
This was born out of a need to react to changes in objects. It is implemented as a function generating a mixin so that you can extend classes you otherwise have no control over. In my initial case it is used with the FormControl from angular in order to observe when it is touched or similar.

Unfortunately it has a pretty bad weakness because it relies on Proxy, which does not go well with the new private properties in JS. The ones with #, they crash in a very confusing way for the user. I have some plans on implementing a custom error message when this happens, but it is a bit advanced and for my own usecase atm I really can't be bothered.

See https://github.com/tc39/proposal-class-fields/issues/106 and https://github.com/littledan/proposal-proxy-transparent and https://github.com/rdking/proposal-proxy-transparency/issues

That this bug is allowed to exist is a bit jarring to say the least. (And yes, as far as I can see, I would classify this as a bug in ES)

BREAKING CHANGE: We are now implementing the actual code. Before there was none. Now there is code.