fix: Resolve self-referencing type bug
The return value of observableForProperty has this[keyof this] which means it is able to watch all properties, which includes the private ones. If they are not visible to the outside it means the class itself would have a different list of variables available than if you watched the instaciated class from outside, meaning the two would have different signatures as one would be able to see the private variable and the other not. As far as I have understood at least. So I will just make everything public, seems easier.

I did contemplate trying to only make keyof AnyClass be the keys that can be used, but it maikes it harder to use the interface and any class that inherits this one will not be able to use the observableForProperty for its properties. Using this is much more expansive and powerrful so I will stick to it.

Also I can now export the type definitions, since everything is public. So I do that as well. Because why not.